IMX7-32 board: phytec: mx7d_phyboard: ensure that the PHY for ETH2 is reset
POR_B reset signal to the PHY is deasserted at the same time that the rail
powering the PHY is turned on, so sometimes the PHY is not reset. Assert the
reset GPIO on boot to ensure that the PHY is reset properly.

Signed-off-by: Anna Horstmann <ahorstmann@phytec.com>
(cherry picked from commit 99aa609eb3d52719667e33a10b5bdae17ae128af)