ARM: OMAP4+: Add support for dynamically selecting OPPs
Recent Kconfig approach for selecting OPPs forces a constraint that a
specific voltage rail should be at the same OPP on all boards supported
by a single defconfig. It can be expected that different paper spins of
a SoC can have different definitions for OPP and can have their own
constraints on the boot up OPP for each voltage rail. In order to have
this flexibility, add support for dynamically selecting the OPP voltage
based on the board to handle any such exceptions or else let use the
Kconfig option from the user.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>