soc: keystone_serdes: update gbe serdes configurations
The updates in this patch mainly concerns improving the serdes
lane stability in the SerDes used by gbe.

After the updates in this patch, the configurations are now
divided mainly in two parts, namely, a common configuration
which is done regardless a SerDes lane is actually used or
not, and a lane specific configuration which is done when a
lane is actually used by, for example, a network interface.

The common configuration is performed by the SerDes provider
initialization, at driver probe. The lane specific configuration
is invoked by the corresponding peripheral driver, for example
network switch driver, if the interface enabled.

The updates have been approved for release by 3rd party under
BSD license.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>