board: ti: am57xx-evm: add newline after error message
Add a newline after touchscreen detection failure
message. Without this, it mixes up with the messages
that follow.

While at it, remove the exclamation after printing
the error code since that gives an impression that
the exclamation is part of the code. The exclamation
if at all desired, should be placed after "failed".

Signed-off-by: Sekhar Nori <nsekhar@ti.com>