dma: keystone_nav: use appropriate Host Descriptor field for dest port
In the K2HK ethss subsystem, the dest port number of a TX packet
is put in the pkt_info field of the TX Host Descriptor (HD) while
in K2E, K2L and K2G, the dest port number should be put in the
tag_info field.

This patch adds the dest_port_info field to the struct pktdma_cfg.
Value of this field indicats which Host Descriptor field,
tag_info or pkt_info, should be used to configure the
destination port number of a TX packet.

The knav_send() will configure appropriate HD field depending on the
dest_port_info in the pktdma parameter passed in.

Prior to this patch, the HD field swinfo[2] was used for that
purpose. Even if that worked, it worked with some subtle errors,
such as packets are actually sent out on all active ports.

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>
Signed-off-by: WingMan Kwok <w-kwok2@ti.com>