dma: keystone_nav: update queue mgr and pkt dma config api
Prior to this patch, queue manager and pkt dma configuration APIs
are designed for the configuration of one queue manager and the
1G cpsw pkt dma only. This patch updates the APIs by allowing the
passing in of additional parameters. The updates facilitate the
reuse of existing code to configure another queue manager that may
exist on some keystone2 platforms. It also makes it more flexible
to configure other peripheral's pkt dma on such platforms.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>