arm: dts: am57xx-phycore-som: limit QSPI to DIO-SPI mode
On revision 1428.3 of the SOM PCB, the QSPI_D[2:3] signals
are not pulled up to VDD_3V3 with external pull-ups.

This is required due to the silicon erratum i916, which
affects both SR1.1 and SR2.0. This erratum notes that the
internal pull-ups are not always enough when a QSPI device
muxes D3 with the HOLD function. Due to the pull-up being
weak, the D3 signal is sometimes still in a low (HOLD) state
when the next command transmission begins unless the device's
OTP/NV registers are programmed to always operate in QSPI mode.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>