arm: dts: am57xx-phytec-: restructure dts organization
We are switching to a model of enabling the core components
and disabling the expansion boards in our SOM and CB dtsi
files.

Expansion board features will be enabled in the applicable dtsi.
Additionally, SOM-specific build configurations will now
disable the unused features and am57xx-phycore-som.dtsi will
remain as the superset of all available features for the
phyCORE-AM572x SOM.

The dts that generates the dtb image for the phyCORE-AM57xx kit
(am57xx-phycore-rdk.dts) will gather all of the components together,
but no longer enable or disable any features itself.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>