board: phytec: am572x_phycore_rdk: Set tps659038 PMIC GPIO7 pad mux value to POWERHOLD
The GPIO7 pad mux should be programmed to POWERHOLD value
as per board design. In cases where the PMIC is shut off the
mux is set to GPIO7 mode. So during initialization to be on the
safer side set the mode to POWERHOLD.

Based on commit 385d3632ba63f9a60047fa9f4f0a7d9629bc0613
	am57xx: Set tps659038 PMIC GPIO7 pad mux value to POWERHOLD
	Author: Keerthy <j-keerthy@ti.com>

Signed-off-by: Russell Robinson <rrobinson@phytec.com>