board: phytec: am572x_phycore_rdk: Register the MMC controllers the same in SPL and u-boot
To keep a consistent MMC device mapping in SPL and in u-boot, let's
register the MMC controllers the same way in u-boot and in the SPL.
In terms of boot time, it doesn't hurt to register more controllers than
needed because the MMC device is initialized only prior being accessed for
the first time.

Having the same device mapping in SPL and u-boot allows us to use the
environment in SPL whatever the MMC boot device.

Based on commit 3eb68c53ebd8aee9ed2b9b58761856a366c2764f
	ti: boot: Register the MMC controllers in SPL in the same way as in u-boot
	Author: Jean-Jacques Hiblot <jjhiblot@ti.com>

Signed-off-by: Russell Robinson <rrobinson@phytec.com>