board: phytec: am572x_phycore_rdk: add override_board env variable
Use 'setenv override_board 1' on the target to allow the
persistent use of a manual board_name configuration in
the environment. If the variable is not set, board_name
will be set based on this function call, meaning it
will be either am572x_phycore_rdk or the appropriate
board_name according to EEPROM data.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>